enrich events with information from blocks
we need the timestamp for each event, which is only stored in the block.

So, after fetching events, we fetch each block by block number that contains an
event.

Strictly speaking we could have one or multiple chain reorgs while fetching
those blocks, so we compare the block hash with the one stored in the event and
bail out if they don't match.

I had already been introducing a new table 'blocks', which would hold some basic
information about blocks. But then decided against it, cause it's a bit easier
to implement this way, i.e. with a single table.

Also, we now use psycopg2.extras.RealDictCursor. This makes psycopg2 give us
results as a dictionary (instead of a tuple).