make runsync more robust
we catch errors in the outer code and are able to handle temporary outages of
postgresql and geth/parity that way.

The disadvantage is that it's now also more robust against configuration errors,
i.e.. it may loop forever if the configuration is wrong, e.g. by setting an
illegal hostname in PGHOST.