schedule getLogs call
For some getLogs calls we can't be sure, which state of the blockchain they
correspond to. The chain could have reorged back and forth and there is no way
for us to detect that.

These getLogs calls will be done a second time when the chain has moved on far
enough.

This commit add the functionality to store getLogs calls to be done at a later
time in the database.

We introduce a new table 'scheduled_get_logs', so users will have to (manually)
update their database schema.

This is only the first part. At the moment there is no code that executes these
scheduled calls.

See https://github.com/trustlines-network/py-eth-index/issues/4