always fetch the latest number_fetch_latest_blocks blocks
since we have no way to reliably detect chain reorgs and it's not clear from
eth.getLogs return value what state the return value corresponds to, we always
fetch the latest number_fetch_latest_blocks blocks.

An alternative would have been to call getTransactionReceipt for every
transaction in a block.

We compare the latest block hash now in order to determine if we need to refetch
logs.

This means that we also had to change our database schema a bit. I'll add that
fact to the changelog later.