add eth-tester
We need this for testing interaction with the blockchain. This adds eth-tester
together with py-evm as a backend. This introduces lot's of new (indirect)
dependencies (as can be seen in the constraints.txt file)

Currently we don't deploy any of our contracts. Instead we test that eth_getLogs
works. I ran into a regresssion with that function while trying to update to
web3 4.5.0. See https://github.com/ethereum/web3.py/issues/999