get basic blockchain test setup working
This commits adds our own "CurrencyNetwork.sol" contract that is being used for
testing. We also commit the compiled contracts.json in order to not have to
install solc.

If the contract is changed, it can be recompiled with 'populus compile'. Though
please note that populus won't install at the moment. Please use a dedicated
virtualenv for populus with the relay server's constraint file to install it.