handle chain reorgs by refetching events
we consider events/blocks from the chain final if they had
Synchronizer.required_confirmations confirmations. When the chain moves forward,
we fetch events for all blocks that were not considered final till the end of
the chain.

If blocks come in one by one, this means we're fetching events for each block
required_confirmations times. This may become something we have to optimize.

In that case take a look at the other solution I proposed in
https://github.com/trustlines-network/py-eth-index/pull/28 (exclude the last
commit), which would only fetch events for each block twice at most. The
disadvantage being that we may have some inconsistencies in the last blocks and
the chain reorg detection would still have to implemented.

Anyway, this solution here is a bit rude, but it is much simpler.

see issue https://github.com/trustlines-network/py-eth-index/issues/4 and
https://github.com/trustlines-network/py-eth-index/pull/28 for my previous
attempt.