replace add_some_tranfer_events with event_emitter fixture
add_some_tranfer_events was a 'side-effect only' fixture, that addes some events
to the blockchain. This was a bit confusing since in the test function this
fixture wasn't used.

Also adds ethereum_tester as a field to TestEnv. This is meant as a preparation
to write tests for the chain reorg handling.

fixes https://github.com/trustlines-network/py-eth-index/issues/36